Cleaning up 'frames' package

+++ jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/api/WebSocket.java
@@ -23,19 +23,19 @@ public class WebSocket
      */
     private final static byte[] MAGIC;
 
-    public final static int CLOSE_NORMAL = 1000;
-    public final static int CLOSE_SHUTDOWN = 1001;
-    public final static int CLOSE_PROTOCOL = 1002;
-    public final static int CLOSE_BAD_DATA = 1003;
-    public final static int CLOSE_UNDEFINED = 1004;
-    public final static int CLOSE_NO_CODE = 1005;
-    public final static int CLOSE_NO_CLOSE = 1006;
-    public final static int CLOSE_BAD_PAYLOAD = 1007;
-    public final static int CLOSE_POLICY_VIOLATION = 1008;
-    public final static int CLOSE_MESSAGE_TOO_LARGE = 1009;
-    public final static int CLOSE_REQUIRED_EXTENSION = 1010;
-    public final static int CLOSE_SERVER_ERROR = 1011;
-    public final static int CLOSE_FAILED_TLS_HANDSHAKE = 1015;
+    public final static short CLOSE_NORMAL = 1000;
+    public final static short CLOSE_SHUTDOWN = 1001;
+    public final static short CLOSE_PROTOCOL = 1002;
+    public final static short CLOSE_BAD_DATA = 1003;
+    public final static short CLOSE_UNDEFINED = 1004;
+    public final static short CLOSE_NO_CODE = 1005;
+    public final static short CLOSE_NO_CLOSE = 1006;
+    public final static short CLOSE_BAD_PAYLOAD = 1007;
+    public final static short CLOSE_POLICY_VIOLATION = 1008;
+    public final static short CLOSE_MESSAGE_TOO_LARGE = 1009;
+    public final static short CLOSE_REQUIRED_EXTENSION = 1010;
+    public final static short CLOSE_SERVER_ERROR = 1011;
+    public final static short CLOSE_FAILED_TLS_HANDSHAKE = 1015;
 
     static
     {
